requirements.txt
Flask
python-pptx
openai
requests
python-dotenv
gunicorn
werkzeug
